Authorization for edit and delete

+++ Back/routes/posts.js
@@ -39,8 +39,10 @@ router.post(
         const post = new Post({
             title: req.body.title,
             content: req.body.content,
-            imagePath: url + "/images/" + req.file.filename
+            imagePath: url + "/images/" + req.file.filename,
+            creator: req.userData.userId
         });
+
         post.save().then(createdPost => {
             res.status(201).json({
                 message: "Post added successfully",
@@ -67,11 +69,17 @@ router.put(
             _id: req.body.id,
             title: req.body.title,
             content: req.body.content,
-            imagePath: imagePath
+            imagePath: imagePath,
+            creator: req.userData.userId
         });
-        console.log(post);
-        Post.updateOne({ _id: req.params.id }, post).then(result => {
-            res.status(200).json({ message: "Update successful!" });
+        Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+            if (result.modifiedCount > 0) {
+                res.status(200).json({ message: "Update successful!" });
+            }
+            else {
+                res.status(401).json({ message: "Not Authorized!" });
+            }
+
         });
     }
 );
@@ -109,9 +117,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id }).then(result => {
-        console.log(result);
-        res.status(200).json({ message: "Post deleted!" });
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.deletedCount > 0) {
+            res.status(200).json({ message: "Deletion successful!" });
+        }
+        else {
+            res.status(401).json({ message: "Not Authorized!" });
+        }
     });
 });
